Delete VM when preempted
Adds `--instance-termination-action=DELETE` to delete instance when
preempted. Without this, when preemptible VM is interrupted it
leaves the instance in shutdown state with disk still incurring
cost.

To use that flag I also had to switch to spot VM provisioning, which
costs the same, have the same base behavior and a few more features.

Spot VMs unlike preemptible ones do not have a maximum runtime
length, so I've picked 3d max runtime as that's already used in the
script as "max workflow runtime".